requirements.txt
Jinja2==3.1.6
MarkupSafe==3.0.2
PySide6_Essentials==6.7.0
PyYAML==6.0.2
certifi==2025.7.14
charset-normalizer==3.4.2
contourpy==1.3.3
cycler==0.12.1
filelock==3.18.0
fonttools==4.59.0
fsspec==2025.7.0
idna==3.10
kiwisolver==1.4.8
matplotlib==3.10.5
mpmath==1.3.0
networkx==3.5
numpy==1.26.4
opencv-python==4.7.0.68
packaging==25.0
pandas==2.3.1
pillow==11.3.0
psutil==7.0.0
py-cpuinfo==9.0.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.4
scipy==1.16.1
shiboken6==6.7.0
six==1.17.0
sympy==1.14.0
torch==2.2.2
torchvision==0.17.2
tqdm==4.67.1
typing_extensions==4.14.1
tzdata==2025.2
ultralytics-thop==2.0.14
ultralytics==8.3.173
urllib3==2.5.0
